First Challenge done

+++ Pages/Index.cshtml.cs
@@ -42,7 +42,7 @@ public class IndexModel : PageModel
                     Id = reader.GetInt32(0),
                     Date = DateTime.Parse(reader.GetString(1), 
                         CultureInfo.CurrentUICulture.DateTimeFormat),
-                    Quantity = reader.GetInt32(2)
+                    Quantity = reader.GetDouble(2)
                 }); ;
             }
